add sync test --cp 🧪 ..
